+++ example/example.js
@@ -18,6 +18,7 @@ const metadata = {
   language: 'en',
   description: 'A test book.',
   contents: 'Chapters',
+  showContents: true,
   source: 'http://www.kcartlidge.com',
   images: ['test/hat.png'],
 };